GM commands /cart and /cg now read from the list_cart_cg.sql file
The sql file is configurable by the admin changing in the isactive field from 0 to 1 to give that item to the gm who uses either command.   

0 = dont use, 1 = allow

Some items in sql file is not in game yet, but are listed in stb, thus was included.

Thanks to lmame for all the work on the code to get this to work!